extract `IdleWrapper`
